requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
librosa==0.10.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0
joblib==1.3.2
python-dotenv==1.0.0
soundfile==0.12.1
scipy==1.11.2
imbalanced-learn==0.11.0
matplotlib==3.7.2
seaborn==0.12.2
torch==2.0.1
torchaudio==2.0.2
Pillow==10.0.0
audioread==3.0.0
